Criação do controller dos chamados

+++ calltasks/src/main/java/br/com/calltasks/model/Chamado.java
@@ -16,7 +16,7 @@ public class Chamado {
 	@Id
 	@Column(name = "codigo_chamado", nullable = false)
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int codigoChamado;
+	private long codigoChamado;
 
 	@Column(name = "dataAbertura_chamado", nullable = false)
 	private java.sql.Timestamp dataAberturaChamado;
